Playing with signals
