+++ src/components/polyline.rs
@@ -1,9 +1,10 @@
-use crate::components::{LeafletMapContext, LeafletOverlayContainerContext};
-use leaflet::{to_lat_lng_array, PolylineOptions};
+use leaflet::{PolylineOptions, to_lat_lng_array};
 use leptos::*;
-use crate::{FillRule, LineCap, LineJoin, MaybeSignalOption};
+use crate::components::context::{extend_context_with_overlay, LeafletMapContext, LeafletOverlayContainerContext, update_overlay_context};
+use crate::MaybeSignalOption;
+use crate::components::path_options::{FillRule, LineCap, LineJoin};
 
-use super::{extend_context_with_overlay, update_overlay_context, Position};
+use super::Position;
 
 #[component(transparent)]
 pub fn Polyline(
